Cleanup of agent cruft. needs testing/logs. this has been a long time coming and should fix multiple issues at startup not finding agents. (which breaks traveler)

+++ Questor.Modules/Actions/Arm.cs
@@ -490,7 +490,7 @@ namespace Questor.Modules.Actions
                 }
 
                 Logging.Log("Arm.MoveDronesToDroneBay", "Dronebay details: Capacity [" + Drones.DroneBay.Capacity + "] UsedCapacity [" + Drones.DroneBay.UsedCapacity + "]", Logging.White);
-                if (Drones.DroneBay.Capacity == Drones.DroneBay.UsedCapacity)
+                if ((int)Drones.DroneBay.Capacity == (int)Drones.DroneBay.UsedCapacity)
                 {
                     if (Drones.DroneBay.Items.Any()
                         && Drones.DroneBay.Items.FirstOrDefault() != null
